re-organized ldl lib
