+++ ldl/ldl.h
