+++ ldl/test/test.c
@@ -2,8 +2,8 @@
 
 #include <stdio.h>
 #include <stdlib.h>
-#include "../include/ldl.h"
-#include "../include/ldl_unittest.h"
+#include "../ldl.h"
+#include "ldl_unittest.h"
 #define ERR 1e-4
 
 float vec_diff_norm(float* x, float* y, int len);
